can delete playlists and songs from playlist

+++ app/src/main/java/com/example/run2thebeat/SelectedPlaylistAdapter.java
@@ -1,4 +1,5 @@
 package com.example.run2thebeat;
+
 import android.graphics.Color;
 import android.view.LayoutInflater;
 import android.view.View;
@@ -11,53 +12,50 @@ import java.util.ArrayList;
 import androidx.annotation.NonNull;
 import androidx.recyclerview.widget.RecyclerView;
 
-public class SelectedPlaylistAdapter extends RecyclerView.Adapter<SelectedPlaylistAdapter.SongViewHolder>{
+public class SelectedPlaylistAdapter extends RecyclerView.Adapter<SelectedPlaylistAdapter.SongViewHolder> {
 
     private ArrayList<Song> songList;
     private SelectedPlaylistAdapter.OnDeleteClickListener mDelete;
     public ImageButton imageButton;
 
 
-    public interface OnDeleteClickListener{
+    public interface OnDeleteClickListener {
         void onDeleteClick(int position);
 
     }
 
-    public void setOnDeleteClickListener(SelectedPlaylistAdapter.OnDeleteClickListener listener){
+    public void setOnDeleteClickListener(SelectedPlaylistAdapter.OnDeleteClickListener listener) {
         mDelete = listener;
     }
 
 
-    public static class SongViewHolder extends RecyclerView.ViewHolder{
+    public static class SongViewHolder extends RecyclerView.ViewHolder {
         public TextView songName;
         public TextView artist;
         public ImageButton deleteSongButton;
-        public SongViewHolder(@NonNull final View item , final SelectedPlaylistAdapter.OnDeleteClickListener deleteListener){
+
+        public SongViewHolder(@NonNull final View item, final SelectedPlaylistAdapter.OnDeleteClickListener deleteListener) {
             super(item);
             songName = item.findViewById(R.id.song_title);
             artist = item.findViewById(R.id.song_artist);
             deleteSongButton = item.findViewById(R.id.icon_delete_song);
-            item.setOnLongClickListener(new View.OnLongClickListener() {
+
+            deleteSongButton.setOnClickListener(new View.OnClickListener() {
                 @Override
-                public boolean onLongClick(View v) {
-                    final int position = getAdapterPosition();
-                    if (position != RecyclerView.NO_POSITION) {
-                        deleteSongButton.setVisibility(View.VISIBLE);
-                        deleteSongButton.setOnClickListener(new View.OnClickListener() {
-                            @Override
-                            public void onClick(View v) {
-                                deleteListener.onDeleteClick(position);
-                            }
-                        });
+                public void onClick(View v) {
+                    if(deleteListener != null){
+                        int position = getAdapterPosition();
+                        if (position != RecyclerView.NO_POSITION) {
+                            deleteListener.onDeleteClick(position);
+                        }
                     }
-                return false;
                 }
             });
         }
     }
 
 
-    public SelectedPlaylistAdapter(ArrayList<Song> myPlayList){
+    public SelectedPlaylistAdapter(ArrayList<Song> myPlayList) {
         songList = myPlayList;
     }
 
@@ -67,7 +65,7 @@ public class SelectedPlaylistAdapter extends RecyclerView.Adapter<SelectedPlayli
     public SelectedPlaylistAdapter.SongViewHolder onCreateViewHolder(@NonNull ViewGroup parent, int viewType) {
         View v = LayoutInflater.from(parent.getContext()).inflate(R.layout.song, parent, false);
 
-        return new SelectedPlaylistAdapter.SongViewHolder(v,mDelete);
+        return new SelectedPlaylistAdapter.SongViewHolder(v, mDelete);
     }
 
     @Override
